combine certificate
